Refactored tests for backend
